Remove duplicated java source
